refactor(app/profile/worker): rename register address
